feat(): i16_t , WIP, 1 test fail !

+++ src/app/assets/examples/for-loop.cpp
@@ -1,9 +1,9 @@
 // For loop
 // Select the whole code to see the conditional structures as a graph
 
-double sum   = 0;
-double count = 5;
-for(double i = 1; i <= count; i = i+1)
+int sum   = 0;
+int count = 5;
+for(int i = 1; i <= count; i = i+1 )
 {
     sum = sum + i;
 }
