+++ src/app/assets/examples/if-else.cpp
@@ -2,8 +2,8 @@
 // Conditional Structures
 // Select the whole code to see the conditional structures as a graph
 
-double a = 5;
-double b = 4;
+int a = 5;
+int b = 4;
 string msg;
 if ( a > b ) {
     msg = "a > b";
