+++ src/app/assets/examples/multi-instructions.cpp
@@ -2,6 +2,6 @@
 // Multiple instructions
 // select the 3 lines above to see the graph:
 
-double f = pow(2, 2) - 1;
-double g = sqrt(64) + 3;
-double h = f + 7 * g;
+int f = pow(2, 2) - 1;
+int g = sqrt(64) + 3;
+int h = f + 7 * g;
